+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [0.1.5] - 2026-08-09
+
+### Fixed
+
+- Backups can now be set up when a paper contains website links. Tesina still
+  blocks links that could run code when it checks a backup file.
+
 ## [0.1.4] - 2026-08-09
 
 ### Added
@@ -92,7 +99,8 @@ and [Semantic Versioning](https://semver.org/).
 - Improved the reliability of installing updates and showing release notes
   after restart.
 
-[Unreleased]: https://github.com/adominicci/apa/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/adominicci/apa/compare/v0.1.5...HEAD
+[0.1.5]: https://github.com/adominicci/apa/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/adominicci/apa/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/adominicci/apa/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/adominicci/apa/compare/v0.1.1...v0.1.2